Mostly MySQL updates
Users can now submit reports. If two reports match, the killed user is no longer able to submit a report, and their dead-ness is represented on the spies page. The killer is then given the target's target. Users can also see who their target is or if they are dead right on the home screen. To do: admin panel is the big thing. Misc small updates and visual updates as well.